Use typed properties
Since we switched to the PHP 7.4 it makes sense to use
typed properties.

One nice things that comes out of it that we can remove
a few checks that are now performed by the engine.

Enum library had to ensure that no internal property is
accessed before it's initialized. That's something
that's done by the engine since PHP 7.4 if we declare
a property with a type and then access it before it was
initialized.